:pencil: change distribution to library and double

+++ Modeling/Distributions/IDistribution.cs
@@ -2,6 +2,6 @@
 {
     public interface IDistribution
     {
-        public uint Generate();
+        public double Generate();
     }
 }
